+++ examples/resources/commonfate_datastax_organization_selector/resource.tf
@@ -1,0 +1,4 @@
+resource "commonfate_datastax_organization_selector" "demo1" {
+  id          = "example"
+  datastax_organization_id = "12345-12345-12345-12345"
+}