+++ internal/provider.go
@@ -157,6 +157,8 @@ func (p *CommonFateProvider) Resources(_ context.Context) []func() resource.Reso
 		NewOktaGroupSelectorResource,
 		NewOktaGroupAvailabilitiesResource,
 		NewDataStaxIntegrationResource,
+		NewDataStaxOrganizationAvailabilitiesResource,
+		NewDataStaxOrganizationSelectorResource,
 	}
 }
 
@@ -259,4 +261,12 @@ func NewOktaGroupAvailabilitiesResource() resource.Resource {
 
 func NewDataStaxIntegrationResource() resource.Resource {
 	return &datastax.DataStaxIntegrationResource{}
+}
+
+func NewDataStaxOrganizationAvailabilitiesResource() resource.Resource {
+	return &datastax.DataStaxOrganizationAvailabilitiesResource{}
+}
+
+func NewDataStaxOrganizationSelectorResource() resource.Resource {
+	return &datastax.DataStaxOrganizationSelectorResource{}
 }